smart_inventory/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
requests
fpdf
fpdf2==2.8.5
azure-storage-blob
azure-servicebus
certifi
opencensus-ext-azure
python-jose
cryptography
gunicorn
python-multipart
azure-identity
azure-functions
azure-keyvault-secrets
python-jose[cryptography]
pydantic
email-validator
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
passlib[bcrypt]
bcrypt<4.0.0
